requirements.txt
alabaster==0.7.12
appdirs==1.4.4
asgiref==3.5.0
astor==0.8.1
atomicwrites==1.4.0
attrs==21.4.0
autoflake==1.4
awsebcli==3.20.3
Babel==2.9.1
bandit==1.7.2
black==20.8b1
botocore==1.23.54
cement==2.8.2
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.3
coverage==5.5
darglint==1.8.1
distlib==0.3.4
Django==3.2.12
django-phone-field==1.8.1
djangorestframework==3.13.1
doc8==0.8.1
docopt==0.6.2
docutils==0.17.1
eradicate==1.0
filelock==3.6.0
flake8==3.9.2
flake8-2020==1.6.1
flake8-bandit==2.1.2
flake8-broken-line==0.2.1
flake8-bugbear==19.8.0
flake8-builtins==1.5.3
flake8-commas==2.1.0
flake8-comprehensions==3.8.0
flake8-debugger==3.2.1
flake8-docstrings==1.6.0
flake8-eradicate==0.3.0
flake8-fixme==1.1.1
flake8-isort==3.0.1
flake8-logging-format==0.6.0
flake8-mutable==1.2.0
flake8-polyfill==1.0.2
flake8-quotes==2.1.2
flake8-rst-docstrings==0.0.12
flake8-string-format==0.2.3
flake8-variables-names==0.0.3
future==0.16.0
geojson==2.5.0
gitdb==4.0.9
GitPython==3.1.26
gunicorn==20.1.0
idna==3.3
imagesize==1.3.0
iniconfig==1.1.1
install==1.3.5
isort==4.3.21
Jinja2==3.0.3
jmespath==0.10.0
MarkupSafe==2.1.0
mccabe==0.6.1
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.9.0
pbr==5.8.1
pep8-naming==0.9.1
pipreqs==0.4.11
platformdirs==2.5.0
pluggy==1.0.0
py==1.11.0
pycodestyle==2.7.0
pydocstyle==5.1.1
pyflakes==2.3.1
Pygments==2.11.2
pyparsing==3.0.7
pypiwin32==223
pytest==6.2.5
pytest-cov==2.12.1
python-dateutil==2.8.2
python-dev-tools==2020.9.10
pytz==2021.3
pyupgrade==2.31.0
pywin32==303
PyYAML==5.4.1
regex==2022.1.18
requests==2.26.0
restructuredtext-lint==1.3.2
semantic-version==2.8.5
Shapely==1.8.1.post1
six==1.14.0
smmap==5.0.0
snowballstemmer==2.2.0
Sphinx==2.4.5
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sqlparse==0.4.2
stevedore==3.5.0
termcolor==1.1.0
testfixtures==6.18.3
tokenize-rt==4.2.1
toml==0.10.2
tox==3.24.5
tox-travis==0.12
typed-ast==1.5.2
typing-extensions==3.10.0.2
tzdata==2021.5
urllib3==1.26.8
virtualenv==20.13.2
wcwidth==0.1.9
wemake-python-styleguide==0.14.1
yarg==0.1.9
